--- game/GameProps.d.ts
@@ -1,2 +1,0 @@
-
-
